+++ CHANGELOG.md
@@ -13,14 +13,17 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - New pause/unpause feature for timers to help with more robust pausing ([#885](https://github.com/excaliburjs/Excalibur/issues/885))
 - New event listening feature to listen to events only `.once(...)` then unsubscribe automatically ([#745](https://github.com/excaliburjs/Excalibur/issues/745))
+- New collision event `postcollision` to indicate if collision resolution occured ([#880](https://github.com/excaliburjs/Excalibur/issues/880))
 
 ### Changed
 
 ### Deprecated
 - `PointerEvent.x` and `PointerEvent.y`, in favor of `PointerEvent.pos` ([#612](https://github.com/excaliburjs/Excalibur/issues/612))
+- `CollisionEvent` has been deprecated in favor of the more clear `PreCollisionEvent` ([#880](https://github.com/excaliburjs/Excalibur/issues/880))
 
 ### Fixed
 - Fixed same instance of color potentially being shared, and thus mutated, between instance actors ([#840](https://github.com/excaliburjs/Excalibur/issues/840))
+- Fixed bug where active and passive type collisions would resolve when they shouldn't when in rigid body mode ([#880]((https://github.com/excaliburjs/Excalibur/issues/880))
 
 <!--------------------------------- DO NOT EDIT BELOW THIS LINE --------------------------------->